+++ geco/graphs/__init__.py
@@ -2,4 +2,4 @@ import geco.graphs.chimera as chimera
 import geco.graphs.pegasus as pegasus
 import geco.graphs.utilities as utilities
 import geco.graphs.lavrov_graph as lavrov_graphs
-import geco.graphs.biqmac_rudy as biqmac_rudy
+import geco.graphs.biqmac as biqmac_rudy
